+++ eved/main.go
@@ -6,7 +6,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/server"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 	"github.com/notional-labs/eve/app"
-	"github.com/notional-labs/eve/eve/cmd"
+	"github.com/notional-labs/eve/eved/cmd"
 )
 
 func main() {
