+++ test_node.sh
@@ -9,27 +9,27 @@ LOGLEVEL="info"
 TRACE="" # "--trace"
 
 
-eve config keyring-backend $KEYRING
-eve config chain-id $CHAINID
-eve config output "json"
+eved config keyring-backend $KEYRING
+eved config chain-id $CHAINID
+eved config output "json"
 
 command -v jq > /dev/null 2>&1 || { echo >&2 "jq not installed. More info: https://stedolan.github.io/jq/download/"; exit 1; }
 
 from_scratch () {
   # remove existing daemon
-  rm -rf ~/.eve/* 
+  rm -rf ~/.eved/* 
 
   # if $KEY exists it should be deleted
   # decorate bright ozone fork gallery riot bus exhaust worth way bone indoor calm squirrel merry zero scheme cotton until shop any excess stage laundry
   # eve1hj5fveer5cjtn4wd6wstzugjfdxzl0xpysfwwn
-  echo "decorate bright ozone fork gallery riot bus exhaust worth way bone indoor calm squirrel merry zero scheme cotton until shop any excess stage laundry" | eve keys add $KEY --keyring-backend $KEYRING --algo $KEYALGO --recover
+  echo "decorate bright ozone fork gallery riot bus exhaust worth way bone indoor calm squirrel merry zero scheme cotton until shop any excess stage laundry" | eved keys add $KEY --keyring-backend $KEYRING --algo $KEYALGO --recover
   # Set moniker and chain-id for Craft
-  eve init $MONIKER --chain-id $CHAINID 
+  eved init $MONIKER --chain-id $CHAINID 
 
   # Function updates the config based on a jq argument as a string
   update_test_genesis () {
     # update_test_genesis '.consensus_params["block"]["max_gas"]="100000000"'
-    cat $HOME/.eve/config/genesis.json | jq "$1" > $HOME/.eve/config/tmp_genesis.json && mv $HOME/.eve/config/tmp_genesis.json $HOME/.eve/config/genesis.json
+    cat $HOME/.eved/config/genesis.json | jq "$1" > $HOME/.eved/config/tmp_genesis.json && mv $HOME/.eved/config/tmp_genesis.json $HOME/.eved/config/genesis.json
   }
 
   # Set gas limit in genesis
@@ -49,24 +49,24 @@ from_scratch () {
 
   # Allocate genesis accounts
   # 10 eve (1 of which is used for validator)
-  eve add-genesis-account $KEY 10000000ueve --keyring-backend $KEYRING
+  eved add-genesis-account $KEY 10000000ueve --keyring-backend $KEYRING
 
   # create gentx with 1 eve
-  eve gentx $KEY 1000000ueve --keyring-backend $KEYRING --chain-id $CHAINID
+  eved gentx $KEY 1000000ueve --keyring-backend $KEYRING --chain-id $CHAINID
 
   # Collect genesis tx
-  eve collect-gentxs
+  eved collect-gentxs
 
   # Run this to ensure everything worked and that the genesis file is setup correctly
-  eve validate-genesis
+  eved validate-genesis
 }
 
 from_scratch
 
 # Opens the RPC endpoint to outside connections
-sed -i '/laddr = "tcp:\/\/127.0.0.1:26657"/c\laddr = "tcp:\/\/0.0.0.0:26657"' ~/.eve/config/config.toml
-sed -i 's/cors_allowed_origins = \[\]/cors_allowed_origins = \["\*"\]/g' ~/.eve/config/config.toml
+sed -i '/laddr = "tcp:\/\/127.0.0.1:26657"/c\laddr = "tcp:\/\/0.0.0.0:26657"' ~/.eved/config/config.toml
+sed -i 's/cors_allowed_origins = \[\]/cors_allowed_origins = \["\*"\]/g' ~/.eved/config/config.toml
 # cors_allowed_origins = []
 
 # # Start the node (remove the --pruning=nothing flag if historical queries are not needed)
-eve start --pruning=nothing  --minimum-gas-prices=0ucraft #--mode validator     
+eved start --pruning=nothing  --minimum-gas-prices=0ucraft #--mode validator
